Refactor tests to use common helpers
