+++ Cargo.toml
@@ -3,6 +3,9 @@ name = "fuzmon"
 version = "0.1.0"
 edition = "2024"
 
+[lib]
+path = "src/lib.rs"
+
 [dependencies]
 nix = { version = "0.28", features = ["ptrace", "process"] }
 addr2line = "0.25"
@@ -20,6 +23,4 @@ zstd = "0.13"
 log = "0.4"
 env_logger = "0.10"
 ctrlc = "3"
-
-[dev-dependencies]
 tempfile = "3"
