+++ tests/fd_events.rs
@@ -36,7 +36,7 @@ sys.stdin.readline()
     let mut child_in = child.stdin.take().expect("stdin");
 
     let logdir = tempdir().expect("logdir");
-    let mut mon = Command::new(env!("CARGO_BIN_EXE_fuzmon"))
+    let mut mon = common::fuzmon_cmd()
         .args([
             "run",
             "-p",
