+++ tests/symbols.rs
@@ -3,6 +3,7 @@ use std::process::{Command, Stdio};
 use tempfile::tempdir;
 
 mod common;
+use fuzmon::test_utils::run_fuzmon_and_check;
 
 fn run_symbol_test(flags: &[&str], expected: &[&str]) {
     let dir = tempdir().expect("tempdir");
@@ -60,7 +61,7 @@ int main() {
 
     let pid = child.id();
     let logdir = tempdir().expect("logdir");
-    common::run_fuzmon_and_check(pid, &logdir, expected);
+    run_fuzmon_and_check(common::fuzmon_cmd().get_program().to_str().unwrap(), pid, &logdir, expected);
 
     unsafe {
         let _ = nix::libc::kill(child.id() as i32, nix::libc::SIGINT);
